pyproject.toml
[project]
name = "bunkabot"
version = "0.1.0"
description = ""
authors = [
    {name = "Hoshi Youmu",email = "hoshi@youmu.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[fast,future]==21.1",
    "fastapi==0.111",
    "uvicorn[standard]==0.30",
    "python-dotenv==1.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
